move tests into component specific test suites
