+++ react-template/src/components/__tests__/header.test.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { shallow, mount } from 'enzyme';
+import { Header } from '../index';
+
+describe('Header Component', () => {
+  it('renders Header without crashing', () => {
+    const div = document.createElement('div');
+    ReactDOM.render(<Header />, div);
+  });
+});
